Mejora sistema de puntuación de reseñas. Ahora se actualiza el promedio en la base de datos. Se modifica el home para mostrar el top de mejores juegos.

+++ gaming_review_aggregate/base/models.py
@@ -17,7 +17,7 @@ class Game(models.Model):
     desarrollador = models.CharField(max_length=150)
     genero = models.CharField(max_length=150)
     plataforma = models.CharField(max_length=150, default="PC")
-    promedio = models.IntegerField(default=0)
+    promedio = models.FloatField(default=0)
 
 
 class Platform(models.Model):
